Export competency matrix to excel done

+++ internal/domain/eduprogscheme.go
@@ -4,12 +4,12 @@ import "time"
 
 type Eduprogscheme struct { //eduprogschemeElement
 	Id                 uint64
-	SemesterNum        uint16
+	SemesterNum        uint64
 	DisciplineId       uint64
 	EduprogId          uint64
 	EduprogcompId      uint64
 	Eduprogcomp        Eduprogcomp
-	CreditsPerSemester uint16
+	CreditsPerSemester uint64
 	CreatedDate        time.Time
 	UpdatedDate        time.Time
 }
